+++ stencil-workspace/src/components/modus-text-input/modus-text-input.tsx
@@ -1,5 +1,5 @@
 // eslint-disable-next-line
-import { Component, Event, EventEmitter, h, Prop } from '@stencil/core';
+import { Component, Event, EventEmitter, h, Method, Prop } from '@stencil/core';
 import { IconSearch } from '../icons/icon-search';
 import { IconClose } from '../icons/icon-close';
 
@@ -11,6 +11,9 @@ import { IconClose } from '../icons/icon-close';
 export class ModusTextInput {
   /** (optional) The input's aria-label. */
   @Prop() ariaLabel: string;
+
+  /** (optional) Sets autofocus on the input. */
+  @Prop() autoFocusInput: boolean;
 
   /** (optional) Whether the input has a clear button. */
   @Prop() clearable = true;
@@ -65,10 +68,16 @@ export class ModusTextInput {
 
   classBySize: Map<string, string> = new Map([
     ['medium', 'medium'],
-    ['large', 'large']
+    ['large', 'large'],
   ]);
 
   textInput: HTMLInputElement;
+
+  /** Focus the input. */
+  @Method()
+  async focusInput(): Promise<void> {
+    this.textInput.focus();
+  }
 
   handleClear(): void {
     this.textInput.value = null;
@@ -86,49 +95,45 @@ export class ModusTextInput {
     const className = `modus-text-input ${this.disabled ? 'disabled' : ''}`;
 
     return (
-      <div
-        aria-disabled={this.disabled}
-        aria-invalid={!!this.errorText}
-        aria-label={this.ariaLabel}
-        aria-placeholder={this.placeholder}
-        aria-readonly={this.readOnly}
-        aria-required={this.required}
-        class={className}>
-        {this.label || this.required
-         ? <div class={'label-container'}>
-             {this.label ? <label>{this.label}</label> : null}{this.required ? <span class="required">*</span> : null}
-           </div>
-         : null
-        }
-        <div
-          class={`input-container ${this.errorText ? 'error' : this.validText ? 'valid' : ''} ${this.classBySize.get(this.size)}`}
-          onClick={() => this.textInput.focus()}>
-            {this.includeSearchIcon ? <IconSearch size="16" /> : null}
-            <input class={`${this.includeSearchIcon ? 'has-left-icon' : ''} ${this.clearable ? 'has-right-icon' : null}`}
-                   disabled={this.disabled}
-                   inputmode={this.inputmode}
-                   maxlength={this.maxLength}
-                   minlength={this.minLength}
-                   onInput={(event) => this.handleOnInput(event)}
-                   placeholder={this.placeholder}
-                   readonly={this.readOnly}
-                   ref={(el) => this.textInput = el as HTMLInputElement}
-                   tabIndex={0}
-                   type={this.type}
-                   value={this.value}/>
-            {this.clearable && !this.readOnly && !!this.value
-             ? <span class="icons clear">
-                  <IconClose onClick={() => this.handleClear()} size="16" />
-                </span>
-             : <span class="icons"></span>
-            }
+      <div aria-disabled={this.disabled} aria-invalid={!!this.errorText} aria-label={this.ariaLabel} aria-placeholder={this.placeholder} aria-readonly={this.readOnly} aria-required={this.required} class={className}>
+        {this.label || this.required ? (
+          <div class={'label-container'}>
+            {this.label ? <label>{this.label}</label> : null}
+            {this.required ? <span class="required">*</span> : null}
+          </div>
+        ) : null}
+        <div class={`input-container ${this.errorText ? 'error' : this.validText ? 'valid' : ''} ${this.classBySize.get(this.size)}`} onClick={() => this.textInput.focus()}>
+          {this.includeSearchIcon ? <IconSearch size="16" /> : null}
+          <input
+            class={`${this.includeSearchIcon ? 'has-left-icon' : ''} ${this.clearable ? 'has-right-icon' : null}`}
+            disabled={this.disabled}
+            inputmode={this.inputmode}
+            maxlength={this.maxLength}
+            minlength={this.minLength}
+            onInput={(event) => this.handleOnInput(event)}
+            placeholder={this.placeholder}
+            readonly={this.readOnly}
+            ref={(el) => (this.textInput = el as HTMLInputElement)}
+            tabIndex={0}
+            type={this.type}
+            value={this.value}
+            autofocus={this.autoFocusInput}
+          />
+          {this.clearable && !this.readOnly && !!this.value ? (
+            <span class="icons clear">
+              <IconClose onClick={() => this.handleClear()} size="16" />
+            </span>
+          ) : (
+            <span class="icons"></span>
+          )}
         </div>
-        {
-          this.errorText ? <label class="sub-text error">{this.errorText}</label> :
-          this.validText ? <label class="sub-text valid">{this.validText}</label> :
-          this.helperText ? <label class="sub-text helper">{this.helperText}</label> :
-          null
-        }
+        {this.errorText ? (
+          <label class="sub-text error">{this.errorText}</label>
+        ) : this.validText ? (
+          <label class="sub-text valid">{this.validText}</label>
+        ) : this.helperText ? (
+          <label class="sub-text helper">{this.helperText}</label>
+        ) : null}
       </div>
     );
   }