+++ CHANGELOG.md
@@ -11,6 +11,10 @@ View all releases at: https://github.com/trimble-oss/modus-web-components/releas
 - Modus Chips active state color is now blue
 - Remove unused Warning button and make Primary button the default
 
+### Added
+
+- Added a prop `auto-focus-input` and a public method `focusInput` to `modus-text-input` component.
+
 ## 0.1.11 - 2022-05-08
 
 ### Fixed